設定の共通化を選択するオプションの追加

+++ app/assets/js/scripts/landing.js
@@ -707,8 +707,9 @@ function dlAsync(login = true){
                     // Build Minecraft process.
                     proc = pb.build()
 
+                    //options.txtを共通化する
                     const copy = require('fs')
-                    if (!copy.existsSync(pb.copyOptionsTxt)){
+                    if (!copy.existsSync(pb.copyOptionsTxt) && ConfigManager.getoptionStandardize() && copy.existsSync(pb.originalOptionsTxt)){
                         copy.copyFile(pb.originalOptionsTxt, pb.copyOptionsTxt, (err) => {
                             if (err) {
                                 console.log(err.stack)
